clean regexp fuzzer
